Revert "refactor: make Request more clean" beacause it makes Request mutable.
This reverts commit c386a43a16d966e277f0a9d6f0c1f9902d1cf795.
